+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- Refuse a line break in a cmd.exe launch argument, which quoting cannot contain ([pm-slack-standup-lvgk](https://github.com/unbraind/pm-slack-standup/blob/main/.agents/pm/issues/pm-slack-standup-lvgk.toon))
+
 ## 2026.9.5 - 2026-09-05
 
 ### Security